# storagenode/pieces: remove buffering from reading/writing and fix io.EOF bug