--- a/storagenode/pieces/readwrite.go
+++ b/storagenode/pieces/readwrite.go
@@ -4,12 +4,9 @@
 package pieces
 
 import (
-	"bufio"
 	"context"
 	"hash"
 	"io"
-
-	"github.com/zeebo/errs"
 
 	"storj.io/storj/pkg/pkcrypto"
 	"storj.io/storj/storage"
@@ -17,7 +14,6 @@ import (
 
 // Writer implements a piece writer that writes content to blob store and calculates a hash.
 type Writer struct {
-	buf  bufio.Writer
 	hash hash.Hash
 	blob storage.BlobWriter
 	size int64
@@ -26,9 +22,8 @@ type Writer struct {
 }
 
 // NewWriter creates a new writer for storage.BlobWriter.
-func NewWriter(blob storage.BlobWriter, bufferSize int) (*Writer, error) {
+func NewWriter(blob storage.BlobWriter) (*Writer, error) {
 	w := &Writer{}
-	w.buf = *bufio.NewWriterSize(blob, bufferSize)
 	w.blob = blob
 	w.hash = pkcrypto.NewHash()
 	return w, nil
@@ -36,7 +31,7 @@ func NewWriter(blob storage.BlobWriter, bufferSize int) (*Writer, error) {
 
 // Write writes data to the blob and calculates the hash.
 func (w *Writer) Write(data []byte) (int, error) {
-	n, err := w.buf.Write(data)
+	n, err := w.blob.Write(data)
 	w.size += int64(n)
 	_, _ = w.hash.Write(data[:n]) // guaranteed not to return an error
 	return n, err
@@ -55,9 +50,6 @@ func (w *Writer) Commit(ctx context.Context) (err error) {
 		return Error.New("already closed")
 	}
 	w.closed = true
-	if err := w.buf.Flush(); err != nil {
-		return Error.Wrap(errs.Combine(err, w.blob.Cancel(ctx)))
-	}
 	return Error.Wrap(w.blob.Commit(ctx))
 }
 
@@ -68,27 +60,24 @@ func (w *Writer) Cancel(ctx context.Context) (err error) {
 		return nil
 	}
 	w.closed = true
-	w.buf.Reset(nil)
 	return Error.Wrap(w.blob.Cancel(ctx))
 }
 
 // Reader implements a piece reader that reads content from blob store.
 type Reader struct {
-	buf  bufio.Reader
 	blob storage.BlobReader
 	pos  int64
 	size int64
 }
 
 // NewReader creates a new reader for storage.BlobReader.
-func NewReader(blob storage.BlobReader, bufferSize int) (*Reader, error) {
+func NewReader(blob storage.BlobReader) (*Reader, error) {
 	size, err := blob.Size()
 	if err != nil {
 		return nil, Error.Wrap(err)
 	}
 
 	reader := &Reader{}
-	reader.buf = *bufio.NewReaderSize(blob, bufferSize)
 	reader.blob = blob
 	reader.size = size
 
@@ -97,7 +86,7 @@ func NewReader(blob storage.BlobReader, bufferSize int) (*Reader, error) {
 
 // Read reads data from the underlying blob, buffering as necessary.
 func (r *Reader) Read(data []byte) (int, error) {
-	n, err := r.buf.Read(data)
+	n, err := r.blob.Read(data)
 	r.pos += int64(n)
 	return n, err
 }
@@ -108,7 +97,6 @@ func (r *Reader) Seek(offset int64, whence int) (int64, error) {
 		return r.pos, nil
 	}
 
-	r.buf.Reset(r.blob)
 	pos, err := r.blob.Seek(offset, whence)
 	r.pos = pos
 	return pos, err
@@ -125,6 +113,5 @@ func (r *Reader) Size() int64 { return r.size }
 
 // Close closes the reader.
 func (r *Reader) Close() error {
-	r.buf.Reset(nil)
 	return Error.Wrap(r.blob.Close())
 }