--- a/storagenode/pieces/store.go
+++ b/storagenode/pieces/store.go
@@ -19,9 +19,7 @@ import (
 )
 
 const (
-	readBufferSize  = 256 * memory.KiB
-	writeBufferSize = 256 * memory.KiB
-	preallocSize    = 4 * memory.MiB
+	preallocSize = 4 * memory.MiB
 )
 
 var (
@@ -98,7 +96,7 @@ func (store *Store) Writer(ctx context.Context, satellite storj.NodeID, pieceID 
 		return nil, Error.Wrap(err)
 	}
 
-	writer, err := NewWriter(blob, writeBufferSize.Int())
+	writer, err := NewWriter(blob)
 	return writer, Error.Wrap(err)
 }
 
@@ -116,7 +114,7 @@ func (store *Store) Reader(ctx context.Context, satellite storj.NodeID, pieceID 
 		return nil, Error.Wrap(err)
 	}
 
-	reader, err := NewReader(blob, readBufferSize.Int())
+	reader, err := NewReader(blob)
 	return reader, Error.Wrap(err)
 }