What: 

* Remove `bufio.Reader` and `bufio.Writer` doesn't seem to have significant impact on reading/writing for large chunk sizes.
* Special case `io.EOF` since most Go code assumes them not to be wrapped.
* Fix bug in endpoint, which assumed `Read` always returns sufficient amount of data.

Please describe the tests:
    (No new functionality; existing tests should cover the changed behavior sufficiently.)
 
Please describe the performance impact:
    Most likely a performance improvement, due to using a read buffer (assuming we do some small or unaligned reads sometimes). Could potentially cause a slight slowdown for read operations in pathological cases, though.

## Code Review Checklist (to be filled out by reviewer)
 - [ ] Does the PR describe what changes are being made?
 - [ ] Does the PR describe why the changes are being made?
 - [ ] Does the code follow [our style guide](https://github.com/storj/docs/blob/master/code/Style.md)?
 - [ ] Does the code follow [our testing guide](https://github.com/storj/docs/blob/master/code/Testing.md)?
 - [ ] Is the PR appropriately sized? (If it could be broken into smaller PRs it should be)
 - [ ] Does the new code have enough tests? (*every* PR should have tests or justification otherwise. Bug-fix PRs especially)
 - [ ] Does the new code have enough documentation that answers "how do I use it?" and "what does it do?"? (both source documentation and [higher level](https://github.com/storj/docs), diagrams?)
 - [ ] Does any documentation need updating?
 - [ ] Do the database access patterns make sense?