@thepaul 2019-07-14
wow ok I guess it does cause problems. should we take out the bufio.Reader entirely? or is there something dumb i'm missing?