@egonelbre 2019-07-15
Two things:
1. we probably should test that the bufio is useful
2. we should figure out what's causing the failure, since adding a buffer somewhere shouldn't break things.