@egonelbre 2019-07-15
PS: One common thing that breaks with adding buffering is wrong code that looks like:
```
n, err := buf.Read(data)
if err != nil {
	return err
}

process(data[:n])
```

Whereas, it should be:

```
n, err := buf.Read(data)
if n > 0 {
	process(data[:n])
}
if err != nil {
	return err
}
```