@thepaul 2019-07-15
Made a ticket to investigate further at V3-2158.