@egonelbre 2019-07-16
So I found a few suspicious read handling:

* mobile/bucket.go:333: resets n to -1 when reader returns io.EOF
* pkg/ranger/file.go:70: resetting of the error code looks suspicious
* storagenode/piecestore/endpoint.go:451: piecereader.Read isn't handling `n`