@egonelbre 2019-07-24
It seems the buffering doesn't matter when writing/reading 512KB at a time:

```
BEFORE:
$ go test -bench . -run XXXX ./storagenode/pieces                       
pkg: storj.io/storj/storagenode/pieces                                  
BenchmarkReadWrite/Write-32                    9         121431378 ns/op
BenchmarkReadWrite/Read-32                   184           6444910 ns/op
PASS                                                                    
ok      storj.io/storj/storagenode/pieces       3.423s                  

AFTER:
$ go test -bench . -run XXXX ./storagenode/pieces                       
goos: windows                                                           
goarch: amd64                                                           
pkg: storj.io/storj/storagenode/pieces                                  
BenchmarkReadWrite/Write-32                    9         126349222 ns/op
BenchmarkReadWrite/Read-32                   193           6145245 ns/op
```