@egonelbre 2019-07-24
The grpc buffer sizes seem appropriate.