@mniewrzal 2019-07-24
hmm its 32KB according to godoc and our buffer was 256KiB before removing.