@egonelbre 2019-07-24
The grpc buffer size is different from the chunk size we transfer.